Created seperate data scraping workflow. Added recon output for scraping workflow
